+++ examples/cpp/PineAPPL.hpp
